requirements.txt
crc16==0.1.1
Django==2.2.5
django-extensions==2.2.1
djangorestframework==3.10.3
pytz==2019.2
redis==3.3.8
six==1.12.0
sqlparse==0.3.0
